+++ test/index.js
@@ -1,3 +1,3 @@
 import parse from '../build/index.js';
 
-parse('sladd --dir=test/output --include=test/sample');
+parse('sladd --dir=test/output --include=test/sample --showchanges=true');