+++ test/sample/export/sample1.js
@@ -1,1 +1,2 @@
 export { sample1 } from './sample1';
+export { index } from '.';