+++ test/sample/import/sample1.js
@@ -3,3 +3,4 @@ import * as sample2 from './sample2.js';
 import * as importImport2 from '../import/sample2.js';
 import * as sampleExport1 from '../export/sample1.js';
 import * as sampleExport2 from '../../output/import/sample2.js';
+import * as index from '.'